// generator/template/src/types/api.d.ts
export interface definitions {
// ...
  ListClassMemberRequest: {
    /**
     * 班级名模糊查询
     */
    classNameLike?: string;
    /**
     * 班级Id
     */
    classRoom_uuid?: string;
    /**
     * 模糊查询（账号或姓名）
     */
    keyWord?: string;
    /**
     * 查询.页号
     */
    pageNo?: number;
    /**
     * 查询.页大小
     */
    pageSize?: number;
    /**
     * 教师Id
     */
    teacherId?: string;
    /**
     * 用户Id
     */
    userId?: string;
  };
// ...
  QueryUserRequest: {
    /**
     * 查询.页号
     */
    pageNo?: number;
    /**
     * 查询.页大小
     */
    pageSize?: number;
    queryParam?: definitions["QueryUserParam"];
  };
// ...
  CreateOrUpdateClassMemberRequest: {
    /**
     * 班级Id
     */
    classId: string;
    /**
     * 是否教师
     */
    isTeacher?: string;
    /**
     * 用户名称
     */
    name?: string;
    /**
     * 学员Id
     */
    userId: string;
  };
// ...
  LoginUserResponse: {
    /**
     * 用户号
     */
    code?: string;
    disable?: boolean;
    /**
     * 姓名
     */
    name?: string;
    /**
     * 角色(admin;teacher;student)
     */
    role?: string;
    token?: string;
  };
// ...
  AddUserRequest: {
    /**
     * 账号
     */
    code?: string;
    /**
     * 邮箱
     */
    email?: string;
    /**
     * 是否也是教师
     */
    isTeacher?: string;
    /**
     * 注册姓名
     */
    name?: string;
  };
// ...
  User: {
    /**
     * 角色.是否是管理员
     */
    admin?: boolean;
    /**
     * 年龄
     */
    age?: number;
    /**
     * 班级名称
     */
    className?: string;
    /**
     * 代码
     */
    code?: string;
    /**
     * 创建(注册)时间
     */
    createTime?: string;
    /**
     * 是否禁用
     */
    disable?: boolean;
    /**
     * 邮箱
     */
    email?: string;
    /**
     * 个人头像相对地址
     */
    headImagePath?: string;
    /**
     * 个人主页
     */
    homePage?: string;
    /**
     * 最近登录地址
     */
    lastestLoginAddr?: string;
    /**
     * 最近登录IP
     */
    lastestLoginIp?: string;
    /**
     * 最近登录时间
     */
    lastestLoginTime?: string;
    /**
     * 名称
     */
    name?: string;
    /**
     * 昵称
     */
    nickname?: string;
    /**
     * 生日
     */
    personalSignature?: string;
    /**
     * 个人简介
     */
    resume?: string;
    /**
     * 教育水平
     */
    selfIntroduction?: string;
    /**
     * 性别:0:女;1:男
     */
    sex?: number;
    /**
     * 角色.是否是学生
     */
    student?: boolean;
    /**
     * 角色.是否是教师
     */
    teacher?: boolean;
    /**
     * 职业
     */
    title?: string;
  };
// ...
  LoginRequest: {
    cas?: boolean;
    /**
     * 用户号
     */
    code: string;
    /**
     * 密码
     */
    password: string;
  };
// ...
}